+++ GameLauncher/App/Classes/LauncherCore/Proxy/NancyGzipCompression.cs
@@ -1,3 +1,4 @@
+using System.IO;
 using System.IO.Compression;
 using System.Linq;
 using Nancy;
@@ -27,12 +28,19 @@ namespace GameLauncher.App.Classes.Proxy
             response.Headers["Content-Encoding"] = "gzip";
             response.Headers["Connection"] = "close";
 
-            var contents = response.Contents;
+            /* Ask System to Allocate Memory */
+            var content = new MemoryStream();
+            /* Response Contents is now feed into Allocated Memory */
+            response.Contents(content);
+            /* Set Position for data in Allocated Memory */
+            content.Position = 0;
+            /* Read the Contents from Allocated Memory */
             response.Contents = responseStream =>
             {
-                using (var compression = new GZipStream(responseStream, CompressionMode.Compress))
+                using (var gzip = new GZipStream(responseStream, CompressionMode.Compress, true))
                 {
-                    contents(compression);
+                    /* Instead of Feeding content Raw (Which can potentially cause OoM) Lets read it from Allocated Memory */
+                    gzip.Write(content.ToArray(), 0, (int)content.Length);
                 }
             };
         }
